requirements.txt
django==3.2
django-bootstrap-v5
django-jquery
django-import-export
django-bootstrap-datepicker-plus
django-mathfilters